+++ package.json
@@ -20,7 +20,6 @@
     "@mui/x-date-pickers": "7.11.1",
     "@reduxjs/toolkit": "2.2.7",
     "@tanstack/react-query": "5.51.21",
-    "axios": "1.7.3",
     "dayjs": "1.11.12",
     "firebase": "10.12.4",
     "html2canvas": "1.4.1",
