+++ src/api/firebaseApp.ts
@@ -1,12 +1,10 @@
 import { initializeApp } from 'firebase/app';
 import { getAuth } from 'firebase/auth';
-import { getFirestore } from 'firebase/firestore';
 import { getDatabase } from 'firebase/database';
 import { firebaseConfig } from './firebaseConfig';
 
 const app = initializeApp(firebaseConfig);
 const auth = getAuth(app);
-const db = getFirestore(app);
 const firebaseDB = getDatabase(app);
 
-export { app, auth, db, firebaseDB };
+export { app, auth, firebaseDB };
