+++ src/pages/salaryList/SalaryListPage.tsx
@@ -5,6 +5,7 @@ import Heading from "../../components/Heading/Heading";
 import * as Styled from './SalaryList.style';
 import { useNavigate } from "react-router-dom";
 import NoticeCard from "./NoticeCard";
+import useSalaryDetails from "./useSalaryDetails";
 
 const years = [
   { value: '2021', text: '2021' },
@@ -13,28 +14,29 @@ const years = [
   { value: '2024', text: '2024' },
 ];
 
-const PayData = [
-  {id: 1, title: '05월 급여명세서', date:dayjs('2024-05-25').format('YYYY.MM.DD'), state:'false'},
-  {id: 2, title: '06월 급여명세서', date:dayjs('2024-06-25').format('YYYY.MM.DD'), state:'false'},
-  {id: 3, title: '07월 급여명세서', date:dayjs('2024-07-25').format('YYYY.MM.DD'), state:'true'}
-]
-PayData.sort((a,b) => b.id-a.id)
-
-const firstPayData = PayData[0]
-const originDate = dayjs(firstPayData.date,'YYYY.MM.DD')
-const finalDate = originDate.format('YYYY년 MM월 ')
-const finalDay = originDate.subtract(2,'day').format('DD일')
-
-export default function SalaryList(){
+export default function SalaryListPage(){
   const navigate = useNavigate()
+  const userId = "sajo1234567"
+  const {data, error, isLoading} = useSalaryDetails()
+
+  if (isLoading) {return <div>Loading...</div>}
+  if (error) {return <div>Error: {error.message}</div>}
+
+  const salaryList = data?.salaryDetails[userId] || [] 
+  salaryList.sort((a,b) => b.id-a.id)
+
+  const firstPayData = salaryList[0]
+  const originDate = dayjs(firstPayData.payday,'YYYY.MM.DD')
+  const finalDate = originDate.format('YYYY년 MM월 ')
+  const finalDay = originDate.subtract(2,'day').format('DD일')
 
   const handleApplicationBtn = (id:number) => {
-    if(PayData.find((item) => item.id === id)){
+    if(salaryList.find((item) => item.id === id)){
       navigate(`/salary-detail/${id}`)
-      }else{
-        console.error('급여 명세서가 없습니다.')
-      }
+    }else{
+          console.error('급여 명세서가 없습니다.')
     }
+  }
 
   return(
     <Styled.Salary>
@@ -49,15 +51,15 @@ export default function SalaryList(){
           menuItems={years}
         />
       </Styled.YearSelect>
-        {PayData.map((el)=>
+        {salaryList.map((el)=>
           (<Styled.ListCardBox key={el.id} $state={el.state} 
             onClick={()=>{handleApplicationBtn(el.id)}}>
             <Styled.List>
             <span className="title">{el.title}</span>
-            <span className="date">{el.date}</span>
+            <span className="date">{el.payday}</span>
             </Styled.List>
             <Styled.Btn>
-              {el.state === 'true' ? 
+              {el.state === true ? 
               <Btn round ='true' label='신청가능'/> 
               : 
               <Btn round='true' disabled label='지급완료'/>
