+++ src/pages/salaryList/useSalaryDetails.tsx
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/react-query";
+import { fetchSalaryDetails } from "./api/fetchSalaryInfo";
+
+export default function useSalaryDetails(){
+  return useQuery({
+    queryKey:['salaryInfo'],
+    queryFn:fetchSalaryDetails
+  })
+}
